MySql

+++ 02-JavaWeb/javaweb/project06-MySQL/sql/DDL.sql
@@ -24,10 +24,10 @@ create table user(
 
 -- 2、创建带约束的表
 create table user(
-                     id int primary key comment '唯一标识',
+                     id int unsigned primary key comment '唯一标识',
                      username varchar(20) not null unique comment '用户名',
                      name varchar(20) not null comment '姓名',
-                     age int comment '年龄',
+                     age tinyint unsigned comment '年龄',
                      gender char(1) default '男' comment '性别'
 ) comment '用户信息表';
